+++ packages/af-utils-react-virtual-headless/src/hooks/useScroller/index.ts
@@ -1,7 +1,7 @@
 import useIsomorphicLayoutEffect from "hooks/useIsomorphicLayoutEffect";
-import List, { ScrollElement } from "models/List";
+import VirtualScroller, { ScrollElement } from "models/VirtualScroller";
 
-const useScroller = (model: List, scroller: ScrollElement) =>
+const useScroller = (model: VirtualScroller, scroller: ScrollElement) =>
     useIsomorphicLayoutEffect(() => {
         model.setScroller(scroller);
         return () => model.setScroller(null);