+++ packages/af-utils-react-virtual-list/decs.d.ts
@@ -1,0 +1,1 @@
+declare module "@af-utils/styled";